pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "The backend of Felipe and Wendy site"
authors = ["Felipe Markson <fmarkson@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.6"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.2"}
requests = "^2.25.1"
motor = "^2.3.0"
dnspython = "^2.0.0"
mercadopago = "^1.1.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
